JBIDE-21425 Server adapter wizard: Finish button should be disabled in page which contains connection selection
